+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 9.0.1 (2019-12-03)
+
+* fixed invalid default parameters for calls to `PDOStatement::fetch()`
+
 ## 9.0.0 (2019-11-22)
 
 ### BC breaks